multi: move `NewFundingScriptTree` to tapscript
Move `NewFundingScriptTree` to tapscript package and rename the func to
`NewChannelFundingScriptTree` to indicate it is channel specific.
